refactoring, removed @Nullable
