+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/facade/PersonFacade.java
@@ -1,6 +1,6 @@
 package ca.mikegabelmann.demo2.persistence.facade;
 
-import ca.mikegabelmann.demo2.persistence.facade.dto.PersonAddress;
+import ca.mikegabelmann.demo2.persistence.facade.bean.PersonAddress;
 
 import java.util.Optional;
