+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/facade/bean/PersonAddress.java
@@ -1,8 +1,7 @@
-package ca.mikegabelmann.demo2.persistence.facade.dto;
+package ca.mikegabelmann.demo2.persistence.facade.bean;
 
 import ca.mikegabelmann.demo2.persistence.model.Address;
 import ca.mikegabelmann.demo2.persistence.model.Person;
-import org.springframework.lang.Nullable;
 
 import java.util.Optional;
 
@@ -10,10 +9,8 @@ import java.util.Optional;
 public class PersonAddress {
     private final Person person;
 
-    @Nullable
     private final Address primaryAddress;
 
-    @Nullable
     private final Address secondaryAddress;
