--- persistence/src/main/java/ca/mikegabelmann/demo2/persistence/facade/dto/package-info.java
@@ -1,8 +1,0 @@
-/**
- * Contains Domain Transfer Objects (DTO) or composite objects. Used to make interaction with one or more service easier
- * when associations may not normally exist.
- */
-@NonNullFields
-package ca.mikegabelmann.demo2.persistence.facade.dto;
-
-import org.springframework.lang.NonNullFields;
