+++ persistence/src/main/java/ca/mikegabelmann/demo2/persistence/model/package-info.java
@@ -1,7 +1,4 @@
 /**
  * This package contains all the classes that support mapping between the database and application using JPA.
  */
-@NonNullFields
 package ca.mikegabelmann.demo2.persistence.model;
-
-import org.springframework.lang.NonNullFields;
