+++ shared/src/main/java/ca/mikegabelmann/demo2/dto/package-info.java
@@ -1,0 +1,7 @@
+/**
+ * <p>These objects are used to communicate between processes or networks. They are typically serialized
+ * and deserialized.</p>
+ * <p>They are typically used to hide the underlying structure of the DB or objects they are derived from in the
+ * database domain.</p>
+ */
+package ca.mikegabelmann.demo2.dto;
